+++ pkNX.Structures/VsTrainer/3DS/TrData6XY.cs
@@ -16,7 +16,7 @@ namespace pkNX.Structures
         public override int Item3 { get => BitConverter.ToUInt16(Data, 0x08); set => BitConverter.GetBytes((ushort)value).CopyTo(Data, 0x08); }
         public override int Item4 { get => BitConverter.ToUInt16(Data, 0x0A); set => BitConverter.GetBytes((ushort)value).CopyTo(Data, 0x0A); }
         public override uint AI { get => BitConverter.ToUInt32(Data, 0x0C); set => BitConverter.GetBytes(value).CopyTo(Data, 0x0C); }
-        public override bool Heal { get => Data[0x10] == 1; set => Data[0x10] = (byte)(value ? 1 : 0); }
+        public override bool Heal { get => Data[0x10] == 1; set => Data[0x10] = value ? 1 : 0; }
         public override int Money { get => Data[0x11]; set => Data[0x11] = (byte)value; }
         public override int Gift { get => BitConverter.ToUInt16(Data, 0x12); set => BitConverter.GetBytes((ushort)value).CopyTo(Data, 0x12); }
     }
